CS50-Final-Project/requirements.txt
cs50
Flask
Flask-Session
requests
OS
tempfile
werkzeug.security
helpers